Fix bug

+++ src/commond/ansys/ansys.go
@@ -52,7 +52,10 @@ func (m Module) execAnsys(e common.ExeContext, args []string, cancel <-chan stru
 	}()
 	select {
 	case <-cancel:
-		common.RL.Info(e, "ansys/execAnsys", "Killing process")
+		if ctx.Process == nil {
+			common.RL.Warn(e, "ansys/execAnsys", "Killing: already exited")
+			return nil
+		}
 		err := ctx.Process.Kill()
 		if err != nil {
 			common.RL.Error(e, "ansys/execAnsys", "Killing process: "+err.Error())
